Refactoring to move ContainerItem and ContainerListing.
